[#6033] Check directory existence before calling DeleteDir
Summary:
In test output (org.yb.pgsql.TestPgForeignKeyBatching-output.txt) I saw the following:
```
ts1|pid7165|:30439|http://127.176.157.54:12806 W1013 21:59:13.792821  7604 db_impl.cc:6444] Failed to cleanup dir /tmp/ybtest-tedyu-1602626160365-1726691044/ts-127.176.157.54-30439-1602626160903/yb-data/tserver/data/rocksdb/table-000030af000030008000000000004038/tablet-8b20afb0385c4fdb8347e17960d9a861/archive: IO error (yb/rocksdb/util/env_posix.cc:278): /tmp/ybtest-tedyu-1602626160365-1726691044/ts-127.176.157.54-30439-1602626160903/yb-data/tserver/data/rocksdb/table-000030af000030008000000000004038/tablet-8b20afb0385c4fdb8347e17960d9a861/archive: No such file or directory
ts1|pid7165|:30439|http://127.176.157.54:12806 W1013 21:59:13.792932  7604 db_impl.cc:6450] Failed to cleanup wal dir /tmp/ybtest-tedyu-1602626160365-1726691044/ts-127.176.157.54-30439-1602626160903/yb-data/tserver/data/rocksdb/table-000030af000030008000000000004038/tablet-8b20afb0385c4fdb8347e17960d9a861: IO error (yb/rocksdb/util/env_posix.cc:278): /tmp/ybtest-tedyu-1602626160365-1726691044/ts-127.176.157.54-30439-1602626160903/yb-data/tserver/data/rocksdb/table-000030af000030008000000000004038/tablet-8b20afb0385c4fdb8347e17960d9a861: No such file or directory
```
We should check directory existence before calling DeleteDir() so that unnecessary warnings are avoided.

Test Plan:
Run TestPgForeignKeyBatching#testDeferredInsertBatching and observe that there is no warning in the above form in org.yb.pgsql.TestPgForeignKeyBatching-output.txt

DestroyDB is referenced 116 times in existing tests. The coverage is wide enough.

Reviewers: bogdan, oleg

Reviewed By: oleg

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D9627

+++ src/yb/rocksdb/db/db_impl.cc
@@ -6439,12 +6439,18 @@ Status DestroyDB(const std::string& dbname, const Options& options) {
     }
 
     // ignore case where no archival directory is present.
-    WARN_NOT_OK(env->DeleteDir(archivedir), "Failed to cleanup dir " + archivedir);
-
+    if (env->FileExists(archivedir).ok()) {
+      WARN_NOT_OK(env->DeleteDir(archivedir), "Failed to cleanup dir " + archivedir);
+    }
     WARN_NOT_OK(env->UnlockFile(lock), "Unlock file failed");
     env->CleanupFile(lockname);
-    WARN_NOT_OK(env->DeleteDir(dbname), "Failed to cleanup dir " + dbname);
-    WARN_NOT_OK(env->DeleteDir(soptions.wal_dir), "Failed to cleanup wal dir " + soptions.wal_dir);
+    if (env->FileExists(dbname).ok()) {
+      WARN_NOT_OK(env->DeleteDir(dbname), "Failed to cleanup dir " + dbname);
+    }
+    if (env->FileExists(soptions.wal_dir).ok()) {
+      WARN_NOT_OK(env->DeleteDir(soptions.wal_dir),
+                  "Failed to cleanup wal dir " + soptions.wal_dir);
+    }
   }
   return result;
 }
